(refactor) Rename package from openapi to oas and update imports in ChatContract and ChatController
